# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Integration with PHP Laravel",
        "description": "Integrate Swagger in Laravel application",
        "contact": {
            "email": "afrozsatvilkar2014@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/label": {
            "get": {
                "summary": "read label",
                "description": "user read label",
                "operationId": "b026240c0c513738271bdfbb0257bdc8",
                "requestBody": {},
                "responses": {
                    "201": {
                        "description": "User successfully registered"
                    },
                    "401": {
                        "description": "The email has already been taken"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "put": {
                "summary": "update label",
                "description": "update user label",
                "operationId": "3cb32932ea53c2520e2532b2e16fa1da",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "id",
                                    "labelname"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label updated Sucessfully"
                    },
                    "404": {
                        "description": "Label Name already exists"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "post": {
                "summary": "create label",
                "description": "create user label",
                "operationId": "4630b7a98d31fc6d438dfc071b1b24a6",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "labelname"
                                ],
                                "properties": {
                                    "labelname": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label added Sucessfully"
                    },
                    "404": {
                        "description": "Label Name already exists"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "summary": "delete label",
                "description": "delete user label",
                "operationId": "d84d55c64cd7aa7415520b938794b859",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label successfully deleted"
                    },
                    "404": {
                        "description": "Label Name already exists"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/notelabel": {
            "post": {
                "summary": "add note label",
                "description": "add note label",
                "operationId": "90a1cfe7bb23f424030a61a17d72701a",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "label_id",
                                    "note_id"
                                ],
                                "properties": {
                                    "label_id": {
                                        "type": "integer"
                                    },
                                    "note_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Label note added Sucessfully"
                    },
                    "409": {
                        "description": "Note Already have a label"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            },
            "delete": {
                "summary": "delete note label",
                "description": "delete note label",
                "operationId": "a87db1177a18a48e9b308e3e1ef0c733",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "label_id",
                                    "note_id"
                                ],
                                "properties": {
                                    "label_id": {
                                        "description": "Request",
                                        "type": "integer"
                                    },
                                    "note_id": {
                                        "description": "Request",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Label successfully deleted"
                    },
                    "404": {
                        "description": "Note not found with this label"
                    },
                    "401": {
                        "description": "Invalid authorization token"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Laravel bearer token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}